fix: VT2-56 UserData 도메인 Dto 데이터 관련 필드 Long 으로 통일

+++ src/main/java/eureca/capstone/project/orchestrator/user/dto/response/user_data/AddBuyerDataResponseDto.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @Builder
 public class AddBuyerDataResponseDto {
     private Long userId;
-    private Integer buyerDataMb;
+    private Long buyerDataMb;
 }
